+++ src/integrationTest/groovy/io/github/devatherock/artifactory/controllers/VersionControllerIntegrationSpec.groovy
@@ -1,0 +1,10 @@
+package io.github.devatherock.artifactory.controllers
+
+import io.micronaut.test.extensions.spock.annotation.MicronautTest
+
+/**
+ * Integration test for {@link VersionController}
+ */
+@MicronautTest(propertySources = 'classpath:application-integration.yml', startApplication = false)
+class VersionControllerIntegrationSpec extends VersionControllerSpec {
+}