+++ src/test/groovy/io/github/devatherock/artifactory/controllers/VersionControllerTestSpec.groovy
@@ -1,0 +1,10 @@
+package io.github.devatherock.artifactory.controllers
+
+import io.micronaut.test.extensions.spock.annotation.MicronautTest
+
+/**
+ * Unit test for {@link VersionController}
+ */
+@MicronautTest(startApplication = true)
+class VersionControllerTestSpec extends VersionControllerSpec {
+}